change protected import pattern (backend.py -> _nccl.py) and docs style from PR suggestions
