+++ pylops_mpi/utils/__init__.py
@@ -1,6 +1,5 @@
 # isort: skip_file
 
 # currently dottest create circular dependency with DistributedArray.py
-# from .dottest import * 
+# from .dottest import *
 from .deps import *
-from .backend import *
